requirements_deploy.txt
google-cloud-run
google-cloud-build
google-cloud-storage
google-auth
